Add threads panel
Shows a panel with all threads where the loggedin user is a participant.
We used the rest api to get all threads in the correct org, then filter
to find the ones in which the user is a participant.
Tried to use the cord Thread component as it is, and it was easy enough
to style without having to build the whole thing up.