Add mark as un/read buttons
A fairly dumb and imperfect first implementation

Ideally we actually want to filter when running useLocationData to only
get unread threadIDs, and/or have another variant of setSeen which lets
you specify a location rather than a threadID and let the server handle
figuring out which threads to update.

For now this just loads the first 35 (I think? initial load of 10 +
fetch 25 more) threads and sets the status on them.

As the comment says, mark as unread is also flawed because the thread is
onscreen when you click the button, which means it will become read again
if you don't navigate away within a couple of seconds. It's a start though!

Test Plan:
https://local.cord.com:7307/ mark some threads unread and some channels
read!

Is a bit janky esp as I think it's reloading fonts a lot... I think that
will be cached in 'prod'?

Reviewers: jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/clack/pull/14