Adds ability to edit a message (#19)
Summary :
We've recently added message edit events and props to Cord - so we'll be
using those here to add this functionality.

This diff has a few new changes but also a bit of refactoring to allow
for that. It's sliightly janky and the UI is not like slack - i think
changing the editing composer UI might be a bit challenging for now. A
few things that haven't made it to this PR include:
- The stop-editing modal when moving channels.
- slack-like UI for the editing composer.