Also log sterr when running update script and return logs from endpoint
Maylynn noticed the update action was failing in Github 😢

When we tried to run the underlying script manually, after ssh-ing into
the machine, it turned out that there were local changes to package-lock
and so git was refusing to pull.  This didn't get logged to github and
it was impossible to know what the problem was - maybe this will help?

Test Plan: :eyes:

Reviewers: maylynn-ng

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/clack/pull/21