Use overflow auto instead of scroll
Setting `overflow: scroll` causes the scroll bars to appear
unconditionally, even if the content isn't big enough to actually need
them. Much of the company uses macOS with a trackpad, in which case
macOS doesn't render a "hard" scroll bar gutter, so this is hard to
notice. But for the sanity of those of us who do use a hardware mouse,
we should set `overflow: auto` which displays the scroll bars only if
the content actually needs them.

Test Plan:
The search box and "users in channel" boxes no longer have spurious
scroll bars appearing for both axes. THe "users in channel" modal does
still have a vertical one, as it does need to scroll that axis.

Reviewers: maylynn-ng, yeo-yeo, km-nur

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/clack/pull/41