Fix issues with the new mobile CSS
There are a couple issues with the new CSS on desktop.  On small
channels, there is a bunch of empty space below the content and thread
list.  Also the user avatar ends up floating in the middle of the
screen on wide windows rather than being paired with the activity
indicator.  Fix those.

Before:

After: