Only return 'available' channels
Adds functionality to find out what Cord orgs the current user is in, and
filter the list of returned channels accordingly.  Adds a couple of
new private channels to test

Test Plan:
Console log the list of available channels.  For my user, it does include
'super-secret-channel' which makes sense because I am a member of it (as
are Josh and Adam), as well as all the non-secret orgs, but does not return
'super-fake-secret-channel', because I am not a member of its corresponding
org (in fact that org doesn't even exist).

Reviewers: jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/clack/pull/47