Add organizationId to a couple more spots
This was pointed out on #48 but it was easier to do this on top of #49
in a separate PR.

The bulk of the changes here are a refactor around `ChannelButton`. That
component was intended to be used by the channel buttons, and used a
thread summary to get the unread state. However the "Threads" button
also used it, which doesn't *have* a thread backing it to get a thread
summary -- but we were incidentally/accidentally fetching one anyway.
This is more problematic when we don't have an org ID to pipe in, so
refactor out the thread summary logic into a wrapper.

Test Plan: TS is happy.

Reviewers: yeo-yeo

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/clack/pull/50