Fix the profile details modal
Make a couple changes to the profile details popup window:

Give the modal `pointer-events: none`, which makes it so when the
modal pops up it doesn't fire a mouse leave event on the avatar

Consolidate the two timeouts into one.  This is both simpler and more
correct, since then a mouseLeave + mouseEnter pair of events won't
result in a hide-then-show sequence of the popup.

Add a click handler that just stops propagation, otherwise if you do
anything inside the popup (like try to highlight the name), it closes
the popup.