Mobile view: hide sidebar when clicking on a notification (#91)
Summary:
Just fixing a TODO we left some time ago. Bumping to v1.21.1 as it fixes
one wrinkle with `onClickNotification` for the
`NotificationListLauncher`

Test plan:
On a narrow viewport, click on the notification launcher to open it,
then click on a notification. We correctly navigate to it, and the
sidebar is correctly hidden. (I.e. no behaviour change)