Don't render composer if we don't yet have an org
Clack needs to load the list of channels, which doesn't happen
instantly, but it still tries to render the rest of the page until that
is done. We can't render the toplevel composer though because it doesn't
have an org until the channel list is done loading, and that's now an
error as of getcord/monorepo#7097.

Test Plan:
Insert a delay into loading the channel list. The composer now appears
after that is done, with no error in console.

Reviewers: yeo-yeo

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/clack/pull/99