Move org check further up to Chat
https://github.com/getcord/clack/pull/99 only loaded the Composer once
the Clack orgs were loaded, because the Composer now errors if it's not
supplied an org and there's no org in the token.

Another place that requires this is user.useOrgMembers, also in the
Chat component.  It doesn't throw but it does log an error which can
be annoying/misleading if you're trying to debug something else, because
it similarly is only an issue for a brief moment while Clack loads the
channels from its servers

Moving the org requirement up is a simple check.  I guess it adds a small
delay to loading a channel but I don't think it will be an issue (and
is just on the first load of Clack, not every time you change channels)

Test Plan: Loading Clack locally looks fine, no error

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/clack/pull/100