After login, make sure user is member of clack_all
We used to rely on this being the org in their token to add them, but
Clack uses the new privacy model, so we need to do this by hand. Any
current employees will already be members, so this only matters for new
employees (where I fixed things by hand the one time it's come up so
far).

Test Plan:
Change the code to authenticate me as a brand-new fake user. Can still
log into Clack and end up as a member of `clack_all`.

Reviewers: AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/clack/pull/112