Use bulk mark as seen (#116)
Use the new bulk mark as seen API to mark all messages in a channel
seen, rather than loading threads and then marking them by ID.

Tested by marking a thread from way back in history in `#noise` as
unread, then marking as seen in prod Clack and local Clack.  It did
not mark it as seen in prod Clack, but did mark it as seen in local
Clack.