Fixing layout message (#128)
## Summary

The grid was completely bonkers when there are replies but no reactions.

Now, instead of having reactions and replies in the same div, they are
separate. Also, setting the grid we want in ClackMessage, depending on
whether a message has reactions or replies

## Test plan
![Screenshot 2024-04-03 at
11.07.50.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/4TJIH6mZTWPlG8Bjs63m/888aa819-9d47-4453-84c6-46e11ba771b6.png)

I checked with attachments and it still works.