Bootstrap users/groups on first start (#135)
For the first run experience, add a couple places where we add the
all_channels_holder user and clack_all group if they don't yet exist.
This makes it so you don't have to do anything specific on the Cord
side when running Clack for the first time, it just creates the helper
bits it needs if they're missing.

The clack_all group is created only in the login path, so it doesn't
work if you delete the group and then load the app, but that doesn't
seem super important to support.