Add display of direct messages (#147)
Add the ability to show direct message channels between an ad hoc
group of users.

The data model here is that direct message channels don't exist inside
`all_channels_holder`, but instead are based on being a member of a
group named `dm:USERID1,USERID2`.  The channel fetcher on the backend
finds all the groups the user is a member of that have the `dm:`
prefix and sends them to the client.  Then the client mostly just
treats those as channels like normal except for using a different
icon.

The main code change is to expand the `Channel` structure to include a
name separate from the ID, and then a chunk of code in the channels
provider to fetch the list of channels, look up the names of all the
users in all the DM channels, and set the names of the channels to the
appropriate things.

This doesn't yet include the ability to create a DM channel, so it
only works if you create these groups manually in the command line,
the Cord console, etc.  That will come in a future diff.

I'm also not super pleased with all the places that have to check for
`channel.id.startsWith(DM_CHANNEL_PREFIX)`.  I might add a separate
property in the `Channel` struct for that to make the code clearer in
a future diff.