Add ability to create direct message channels (#149)
Add a button to the channel list that opens a modal dialog for
selecting which users to create a DM with, then calls an updated
server call that creates the channel by creating the appropriate
group in the Cord backend.

The dialog is mostly a duplicate of the UsersInChannelModal, but it
needs to be different enough that it makes sense to make it its own
thing rather than trying to share code.