Fix display of two list-of-users modals
These two modals (which are very similar, and maybe should be
combined) both had a problem where in some cases and in some display
sizes they looked weird.  The commit button could render outside the
bounds of the modal, the header text wasn't visually centered within
the header, and the close box was weirdly aligned with the rest of the
header.  Fix those things.