Add separate threadName channel property (#163)
The display name of a DM channel (the `name` property) omits the
user's name, since all DMs the user can use must have them in it, so
it's just noise.  However, the name of the thread is frequently shown
to other people, such as in email notifications of new messages, and
omitting the viewer means that if Person A sends a DM to Person B,
then Person B will get an email notification with a title such as "New
activity on Person B", which doesn't make sense.

Instead, ship around a separate `threadName` property that does
include the user's name and we use as the thread name, so that it will
say "New activity on Person A, Person B", which looks much more
correct for a DM.