Stop using placeholder channel
Instead of generating a placeholder channel when someone links to
something that doesn't exist, use an explicit `undefined` value and
don't render a Chat component.  The Chat component that got rendered
for a placeholder channel worked, but it sent the messages off into
the ether, which is weird.  Instead, rendering nothing shows something
has gone wrong and you need to pick a channel that exists.