Fetch more threads on scroll (#175)
When scrolling to the top, if there are more threads to be loaded, load
them.

This doesn't mimic the previous behaviour, when we were preserving the
scroll. Now, the scroll container will scroll to the top as soon as the
new threads are loaded. I need to investigate more how to do it.