Add Giphy Integration to Clack
Wanted to be able to attach Gifs to my
messages, and using the new beta components allows
me to add this Giphy Integration!

Discussed with Adam about how to do this and we ended
up with adding an api route to get the api token
from the server and store this within a config context
(incase we have any more that we want to retrieve).

I've applied for a production key from Giphy
(currently we can make a max of 100 api calls per
hour) so hopefully we can get this soon.

Test Plan:
Add GIPHY_API_KEY in .env and ran Clack
locally. Can attach and send gifs!

Reviewers: dmmiller, flooey

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/clack/pull/172