Fix tooltip not showing when hovering the menus (#183)
There's an issue with the z-index of the tooltips, so they're not
showing when hovering the menus. Now they should show.

Also, it doesn't really make sense to have 3 different tooltips for the
different buttons, we can use the same one. That way we also avoid
problems of one still showing while another one is shown, because we'll
reuse the only one when hovering new buttons.